+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 * `penman.tree.Tree.reset_variables()` ([#41][])
 * `--make-variables=FMT` command-line option ([#41][])
+* `penman.surface.AlignmentMarker.from_string()` ([#45][])
 
 ### Removed
 
@@ -13,7 +14,7 @@
 * `FLOAT` and `INTEGER` productions in `penman.lexer.PATTERNS` ([#44][])
 * `penman.codec.PENMANCodec.ATOMS` ([#44][])
 * `penman.codec.PENMANCodec.IDENTIFIERS` ([#44][])
-* `COMMA` and `CARET` productions in `penma.lexer.PATTERNS` ([#43][])
+* `COMMA` and `CARET` productions in `penman.lexer.PATTERNS` ([#43][])
 * `triples` parameter for the following ([#42][]):
   - `penman.interface.decode()`
   - `penman.interface.loads()`
@@ -23,11 +24,16 @@
   - `penman.interface.dump()`
   - `penman.codec.encode()`
   - `penman.codec.decode()`
+* `ALIGNMENT` production in `penman.lexer.PATTERNS` ([#45][])
 
 ### Changed
 
 * Remove support for numeric data detection and casting ([#44][])
 * Allow commas and carets in the `SYMBOL` pattern ([#43][])
+* Allow tildes in the `ROLE` and `SYMBOL` patterns ([#45][])
+* Surface alignments are extracted and formatted in `penman.layout`,
+  not `penman.lexer` or `penman.codec` ([#45][])
+* Tree branches are simple 2-tuples: `(role, target)` ([#45][])
 
 
 ## [v0.8.0][]
@@ -482,3 +488,4 @@ First release with very basic functionality.
 [#42]: https://github.com/goodmami/penman/issues/42
 [#43]: https://github.com/goodmami/penman/issues/43
 [#44]: https://github.com/goodmami/penman/issues/44
+[#45]: https://github.com/goodmami/penman/issues/45